maintain 118

+++ yukicoder/tyama_yukicoder118.c
@@ -4,7 +4,7 @@ int main(){
 	int N,x,i,j,k;
 	for(scanf("%d",&N);N--;)scanf("%d",&x),a[x]++;
 	for(i=1;i<=M-2;i++)for(j=i+1;j<=M-1;j++)for(k=j+1;k<=M;k++){
-		r+=a[i]*a[j]*a[k];
+		r+=(long long)a[i]*a[j]*a[k];
 	}
 	printf("%lld\n",r%1000000007);
 }
